openenv-bank-reconciliation/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
openai>=1.0.0
faker>=20.0.0
rapidfuzz>=3.0.0
python-dotenv>=1.0.0
openenv-core>=0.1.0
